* thread_pthread.c (rb_thread_create_timer_thread):   the type of retrun value of write(2) is ssize_t.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@32246 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ thread_pthread.c
@@ -993,7 +993,7 @@ static int timer_thread_pipe_owner_process;
 void
 rb_thread_wakeup_timer_thread(void)
 {
-    int result;
+    ssize_t result;
 
     /* already opened */
     if (timer_thread_pipe_owner_process == getpid()) {
@@ -1023,7 +1023,7 @@ consume_communication_pipe(void)
 {
     const size_t buff_size = 1024;
     char buff[buff_size];
-    int result;
+    ssize_t result;
   retry:
     result = read(timer_thread_pipe[0], buff, buff_size);
     if (result < 0) {
